+++ src/ext/java/org/opentripplanner/ext/gtfsgraphqlapi/datafetchers/stepImpl.java
@@ -5,8 +5,10 @@ import graphql.schema.DataFetchingEnvironment;
 import org.opentripplanner.ext.gtfsgraphqlapi.generated.GraphQLDataFetchers;
 import org.opentripplanner.ext.gtfsgraphqlapi.generated.GraphQLTypes.GraphQLAbsoluteDirection;
 import org.opentripplanner.ext.gtfsgraphqlapi.generated.GraphQLTypes.GraphQLRelativeDirection;
+import org.opentripplanner.ext.gtfsgraphqlapi.mapping.DirectionMapper;
 import org.opentripplanner.ext.gtfsgraphqlapi.mapping.StreetNoteMapper;
 import org.opentripplanner.model.plan.ElevationProfile.Step;
+import org.opentripplanner.model.plan.RelativeDirection;
 import org.opentripplanner.model.plan.WalkStep;
 import org.opentripplanner.routing.alertpatch.TransitAlert;
 
@@ -15,21 +17,7 @@ public class stepImpl implements GraphQLDataFetchers.GraphQLStep {
   @Override
   public DataFetcher<GraphQLAbsoluteDirection> absoluteDirection() {
     return environment ->
-      getSource(environment)
-        .getAbsoluteDirection()
-        .map(dir ->
-          switch (dir) {
-            case NORTH -> GraphQLAbsoluteDirection.NORTH;
-            case NORTHEAST -> GraphQLAbsoluteDirection.NORTHEAST;
-            case EAST -> GraphQLAbsoluteDirection.EAST;
-            case SOUTHEAST -> GraphQLAbsoluteDirection.SOUTHEAST;
-            case SOUTH -> GraphQLAbsoluteDirection.SOUTH;
-            case SOUTHWEST -> GraphQLAbsoluteDirection.SOUTHWEST;
-            case WEST -> GraphQLAbsoluteDirection.WEST;
-            case NORTHWEST -> GraphQLAbsoluteDirection.NORTHWEST;
-          }
-        )
-        .orElse(null);
+      getSource(environment).getAbsoluteDirection().map(DirectionMapper::map).orElse(null);
   }
 
   @Override
@@ -79,22 +67,7 @@ public class stepImpl implements GraphQLDataFetchers.GraphQLStep {
 
   @Override
   public DataFetcher<GraphQLRelativeDirection> relativeDirection() {
-    return environment ->
-      switch (getSource(environment).getRelativeDirection()) {
-        case DEPART -> GraphQLRelativeDirection.DEPART;
-        case HARD_LEFT -> GraphQLRelativeDirection.HARD_LEFT;
-        case LEFT -> GraphQLRelativeDirection.LEFT;
-        case SLIGHTLY_LEFT -> GraphQLRelativeDirection.SLIGHTLY_LEFT;
-        case CONTINUE -> GraphQLRelativeDirection.CONTINUE;
-        case SLIGHTLY_RIGHT -> GraphQLRelativeDirection.SLIGHTLY_RIGHT;
-        case RIGHT -> GraphQLRelativeDirection.RIGHT;
-        case HARD_RIGHT -> GraphQLRelativeDirection.HARD_RIGHT;
-        case CIRCLE_CLOCKWISE -> GraphQLRelativeDirection.CIRCLE_CLOCKWISE;
-        case CIRCLE_COUNTERCLOCKWISE -> GraphQLRelativeDirection.CIRCLE_COUNTERCLOCKWISE;
-        case ELEVATOR -> GraphQLRelativeDirection.ELEVATOR;
-        case UTURN_LEFT -> GraphQLRelativeDirection.UTURN_LEFT;
-        case UTURN_RIGHT -> GraphQLRelativeDirection.UTURN_RIGHT;
-      };
+    return environment -> DirectionMapper.map(getSource(environment).getRelativeDirection());
   }
 
   @Override