+++ src/main/java/org/opentripplanner/api/mapping/RelativeDirectionMapper.java
@@ -9,35 +9,23 @@ public class RelativeDirectionMapper {
     if (domain == null) {
       return null;
     }
-    switch (domain) {
-      case DEPART:
-        return ApiRelativeDirection.DEPART;
-      case HARD_LEFT:
-        return ApiRelativeDirection.HARD_LEFT;
-      case LEFT:
-        return ApiRelativeDirection.LEFT;
-      case SLIGHTLY_LEFT:
-        return ApiRelativeDirection.SLIGHTLY_LEFT;
-      case CONTINUE:
-        return ApiRelativeDirection.CONTINUE;
-      case SLIGHTLY_RIGHT:
-        return ApiRelativeDirection.SLIGHTLY_RIGHT;
-      case RIGHT:
-        return ApiRelativeDirection.RIGHT;
-      case HARD_RIGHT:
-        return ApiRelativeDirection.HARD_RIGHT;
-      case CIRCLE_CLOCKWISE:
-        return ApiRelativeDirection.CIRCLE_CLOCKWISE;
-      case CIRCLE_COUNTERCLOCKWISE:
-        return ApiRelativeDirection.CIRCLE_COUNTERCLOCKWISE;
-      case ELEVATOR:
-        return ApiRelativeDirection.ELEVATOR;
-      case UTURN_LEFT:
-        return ApiRelativeDirection.UTURN_LEFT;
-      case UTURN_RIGHT:
-        return ApiRelativeDirection.UTURN_RIGHT;
-      default:
-        throw new IllegalArgumentException(domain.toString());
-    }
+    return switch (domain) {
+      case DEPART -> ApiRelativeDirection.DEPART;
+      case HARD_LEFT -> ApiRelativeDirection.HARD_LEFT;
+      case LEFT -> ApiRelativeDirection.LEFT;
+      case SLIGHTLY_LEFT -> ApiRelativeDirection.SLIGHTLY_LEFT;
+      case CONTINUE -> ApiRelativeDirection.CONTINUE;
+      case SLIGHTLY_RIGHT -> ApiRelativeDirection.SLIGHTLY_RIGHT;
+      case RIGHT -> ApiRelativeDirection.RIGHT;
+      case HARD_RIGHT -> ApiRelativeDirection.HARD_RIGHT;
+      case CIRCLE_CLOCKWISE -> ApiRelativeDirection.CIRCLE_CLOCKWISE;
+      case CIRCLE_COUNTERCLOCKWISE -> ApiRelativeDirection.CIRCLE_COUNTERCLOCKWISE;
+      case ELEVATOR -> ApiRelativeDirection.ELEVATOR;
+      case UTURN_LEFT -> ApiRelativeDirection.UTURN_LEFT;
+      case UTURN_RIGHT -> ApiRelativeDirection.UTURN_RIGHT;
+      case ENTER_STATION -> ApiRelativeDirection.ENTER_STATION;
+      case EXIT_STATION -> ApiRelativeDirection.EXIT_STATION;
+      case FOLLOW_SIGNS -> ApiRelativeDirection.FOLLOW_SIGNS;
+    };
   }
 }