+++ src/test/java/org/opentripplanner/test/support/TestTableParser.java
@@ -45,18 +45,7 @@ public class TestTableParser {
     var result = new Object[args.length];
 
     for (int i = 0; i < args.length; i++) {
-      String arg = args[i].toLowerCase(Locale.ROOT);
-      if (arg.matches("-?\\d+")) {
-        result[i] = Integer.valueOf(arg);
-      } else if (arg.matches("-?\\d*\\.\\d+")) {
-        result[i] = Double.valueOf(arg);
-      } else if (arg.matches("t(rue)?|y(es)?|x")) {
-        result[i] = Boolean.TRUE;
-      } else if (arg.matches("f(alse)?|n(o)?|-")) {
-        result[i] = Boolean.FALSE;
-      } else {
-        result[i] = arg;
-      }
+      result[i] = toValue(args[i]);
     }
     return result;
   }
@@ -64,6 +53,23 @@ public class TestTableParser {
   private static String stripComment(String line) {
     int pos = line.indexOf("#");
     return pos < 0 ? line : (pos == 0 ? "" : line.substring(0, pos).trim());
+  }
+
+  private static Object toValue(String arg) {
+    if (arg.matches("-?\\d+")) {
+      return Integer.valueOf(arg);
+    }
+    if (arg.matches("-?\\d*\\.\\d+")) {
+      return Double.valueOf(arg);
+    }
+    String lcArg = arg.toLowerCase(Locale.ROOT);
+    if (lcArg.matches("t(rue)?|y(es)?|x")) {
+      return Boolean.TRUE;
+    }
+    if (lcArg.matches("f(alse)?|n(o)?|-")) {
+      return Boolean.FALSE;
+    }
+    return arg;
   }
 
   @Test